v0.17
I redesigned the flex pcb to use the TS3A5018 chip
for switching instead of the ADG774A.  This supports
a 3v signal, whereas the ADG774A only supported up
to 1.5v.  Also a bunch of minor tweaks to the board
design to make it easier to assemble and test (added
test pads!)

Also made a rigid version of the flex pcb, to help
debug manufacturing defects if necessary.